backend/requirements.txt
flask==2.2.5
flask-socketio==5.3.4
python-socketio==5.9.0
python-engineio==4.8.2
eventlet==0.33.3
gunicorn==20.1.0
